refactor: introduce o.o.plugins package
- gather internal plugins related classes into o.o.plugins
- refactor o.o.p.PluginType enum
