+++ aligner/src/main/java/org/omegat/gui/align/AlignFilePickerController.java
@@ -63,7 +63,7 @@ import org.omegat.util.Preferences;
 import org.omegat.util.StringUtil;
 import org.omegat.util.gui.LanguageComboBoxRenderer;
 import org.omegat.util.gui.StaticUIUtils;
-import org.omegat.util.module.PluginLifecycleManager;
+import org.omegat.plugins.PluginLifecycleManager;
 
 /**
  * Controller for align file picker UI
