+++ src/org/omegat/gui/exttrans/MachineTranslateTextArea.java
@@ -53,7 +53,7 @@ import org.omegat.core.data.SourceTextEntry;
 import org.omegat.core.data.TMXEntry;
 import org.omegat.core.machinetranslators.MachineTranslateError;
 import org.omegat.core.machinetranslators.MachineTranslators;
-import org.omegat.filters2.master.PluginUtils;
+import org.omegat.plugins.PluginUtils;
 import org.omegat.gui.common.EntryInfoSearchThread;
 import org.omegat.gui.common.EntryInfoThreadPane;
 import org.omegat.gui.glossary.GlossaryEntry;
