+++ src/org/omegat/gui/preferences/view/PluginInfoTableModel.java
@@ -31,8 +31,8 @@ import java.util.List;
 
 import javax.swing.table.DefaultTableModel;
 
-import org.omegat.core.data.PluginInformation;
-import org.omegat.filters2.master.PluginUtils;
+import org.omegat.plugins.PluginInformation;
+import org.omegat.plugins.PluginUtils;
 import org.omegat.util.OStrings;
 
 public class PluginInfoTableModel extends DefaultTableModel {
