+++ test/src/org/omegat/plugins/PluginInformationTest.java
@@ -1,29 +1,29 @@
-/**************************************************************************
- OmegaT - Computer Assisted Translation (CAT) tool
-          with fuzzy matching, translation memory, keyword search,
-          glossaries, and translation leveraging into updated projects.
+/*******************************************************************************
+ *  OmegaT - Computer Assisted Translation (CAT) tool
+ *           with fuzzy matching, translation memory, keyword search,
+ *           glossaries, and translation leveraging into updated projects.
+ *
+ *  Copyright (C) 2021-2024 Hiroshi Miura
+ *                Home page: https://www.omegat.org/
+ *                Support center: https://omegat.org/support
+ *
+ *  This file is part of OmegaT.
+ *
+ *  OmegaT is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  OmegaT is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ ******************************************************************************/
 
- Copyright (C) 2021 Hiroshi Miura
-               Home page: https://www.omegat.org/
-               Support center: https://omegat.org/support
-
- This file is part of OmegaT.
-
- OmegaT is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
-
- OmegaT is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <https://www.gnu.org/licenses/>.
- **************************************************************************/
-
-package org.omegat.core.data;
+package org.omegat.plugins;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -38,8 +38,6 @@ import java.nio.file.Files;
 import java.util.jar.Manifest;
 
 import org.junit.Test;
-
-import org.omegat.filters2.master.PluginUtils;
 
 
 /**
@@ -59,7 +57,7 @@ public class PluginInformationTest {
                  .fromManifest(pluginClass, m, mu, PluginInformation.Status.INSTALLED);
          assertEquals("Filters for OmegaT", pluginInformation.getName());
          assertEquals("Example Author", pluginInformation.getAuthor());
-         assertEquals(PluginUtils.PluginType.FILTER, pluginInformation.getCategory());
+         assertEquals(PluginType.FILTER, pluginInformation.getCategory());
          assertEquals("https://example.com", pluginInformation.getLink());
          assertEquals(mu, pluginInformation.getUrl());
          assertFalse(pluginInformation.isBundled());
