+++ test/src/svn/PluginsTest.java
@@ -37,13 +37,14 @@ import java.nio.file.Paths;
 import java.util.Properties;
 
 import org.junit.Test;
-import org.omegat.filters2.master.PluginUtils;
+
+import org.omegat.plugins.PluginLifecycleManager;
 import org.omegat.util.StaticUtils;
 import org.omegat.util.TestPreferencesInitializer;
 
 public class PluginsTest {
 
-    private static final Path PLUGINS_FILE = Paths.get(StaticUtils.installDir(), PluginUtils.PLUGINS_LIST_FILE);
+    private static final Path PLUGINS_FILE = Paths.get(StaticUtils.installDir(), PluginLifecycleManager.PLUGINS_LIST_FILE);
 
     @Test
     public void testPluginsListWhitespace() throws Exception {
